Move the taxidata program and refactor it for use as a library.
