+++ nyctaxi/taxidata/.gitignore
@@ -1,0 +1,2 @@
+/taxidata
+/taxidata.exe
