+++ nyctaxi/taxidata/Makefile
@@ -1,0 +1,7 @@
+.PHONY: build run
+
+build:
+	go build
+
+run:
+	go run main.go
