+++ nyctaxi/taxidata/go.mod
@@ -1,0 +1,18 @@
+module github.com/streamnative/examples/nyctaxi/taxidata
+
+go 1.15
+
+require (
+	github.com/apache/pulsar-client-go v0.3.0
+	github.com/apache/pulsar-client-go/oauth2 v0.0.0-20200825011529-c078454b47b6 // indirect
+	github.com/golang/protobuf v1.4.3 // indirect
+	github.com/google/go-cmp v0.5.4 // indirect
+	github.com/spf13/cobra v0.0.3
+	github.com/spf13/pflag v1.0.3
+	github.com/stretchr/testify v1.5.1 // indirect
+	golang.org/x/net v0.0.0-20201209123823-ac852fbbde11 // indirect
+	golang.org/x/oauth2 v0.0.0-20201208152858-08078c50e5b5 // indirect
+	golang.org/x/sys v0.0.0-20210104204734-6f8348627aad // indirect
+	golang.org/x/text v0.3.4 // indirect
+	google.golang.org/appengine v1.6.7 // indirect
+)
