--- pulsar-flink/go/go.mod
@@ -1,11 +1,0 @@
-module github.com/streamnative/examples/pulsar-flink-taxidata
-
-go 1.15
-
-require (
-	github.com/apache/pulsar-client-go v0.3.0
-	github.com/apache/pulsar-client-go/oauth2 v0.0.0-20200825011529-c078454b47b6 // indirect
-	github.com/spf13/cobra v0.0.3
-	github.com/spf13/viper v1.7.1 // indirect
-	google.golang.org/api v0.40.0 // indirect
-)
